refactor(backend): batch dtos related to user in single file
