+++ backend/src/dto/mod.rs
@@ -1,1 +1,1 @@
-pub mod user_get_dto;
+pub mod user_dto;
